--- site/server/views/CustomDataExplorerPage.tsx
@@ -1,14 +1,0 @@
-import * as React from "react"
-import { DataExplorerPage } from "./DataExplorerPage"
-
-export const CustomDataExplorerPage = () => {
-    return (
-        <DataExplorerPage
-            title="Data explorer"
-            slug="custom-data-explorer"
-            imagePath=""
-            preloads={[]}
-            inlineJs=""
-        />
-    )
-}
